# openapi.yaml
openapi: 3.0.0
info:
  title: CheckChat API
  description: API for CheckChat App
  version: 1.0.0
servers:
  - url: https://hidden-bayou-91236-aef7370094a2.herokuapp.com
    description: Heroku server
  - url: localhost:3000
    description: Local server

paths:
  /conversations:
    get:
      summary: Get all conversations
      responses:
        '200':
          description: An array of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
    put:
      summary: Update a conversation
      description: Updates the details of an existing conversation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Conversation'
      responses:
        '200':
          description: Conversation updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '404':
          description: Conversation not found

  /conversation/{id}:
    get:
      summary: Get a conversation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A single conversation object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '404':
          description: Conversation not found

  /messages:
    get:
      summary: Get all messages
      responses:
        '200':
          description: An array of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
    put:
      summary: Update a message
      description: Updates the details of an existing message.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        '200':
          description: Message updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Message not found

  /message/{id}:
    get:
      summary: Get a message by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A single message object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Message not found

  /messages/conversation/{conversation_id}:
    get:
      summary: Get messages for a specific conversation ID
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: An array of messages for the given conversation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
                      

  /upload/{conversation_id}/{message_id}:
    post:
      summary: Upload a file
      description: Uploads a file to a specific message in a conversation.
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
        - name: message_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: File uploaded successfully
        '400':
          description: No file uploaded

components:
  schemas:
    Message:
      type: object
      properties:
        id:
          type: string
        conversation_id:
          type: string
        # Define other properties of a Message

    Conversation:
      type: object
      properties:
        id:
          type: string
        # Define other properties of a Conversation

              
  /reset:
    get:
      summary: Reset all messages and conversations
      responses:
        '200':
          description: status response
          

components:
  schemas:
    Conversation:
      type: object
      properties:
        id:
          type: integer
        service_provider_name:
          type: string
        customer_name:
          type: string
        state:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Message:
      type: object
      properties:
        id:
          type: integer
        conversation_id:
          type: integer
        sender_type:
          type: string
        message_type:
          type: string
        text:
          type: string
        created_at:
          type: string
          format: date-time
        read_at:
          type: string
          format: date-time
        hidden_at:
          type: string
          format: date-time
  